refactor(extract): move extract_terraform to extractors/terraform.py (verbatim)
Ports the Terraform/HCL extractor out of the 17k-line extract.py into its
own per-language module, continuing the split tracked in #1212.

- extract_terraform and its private _TF_META_HEADS constant moved verbatim
  (git diff --color-moved confirms byte-identity; the existing test_terraform.py
  suite passing is the behavior-preservation proof).
- extract.py keeps the facade re-export so all existing importers are unchanged.
- Registered in extractors/__init__.LANGUAGE_EXTRACTORS (alphabetical).
- Added tests/test_extractors_registry.py with module/facade/registry identity
  checks; MIGRATION.md status table updated.
